HLVL 0.109.0
Depth buffering

Changes:
- Added method to vulkan factory for getting depth resources
- Updated renderer to account for depth iamge
- Updated pipeline to account for depth stencil

+++ src/core/include/resource.hpp
@@ -58,7 +58,7 @@ class Resource : public ResourceProxy {
       memcpy((char *)memoryMap + offsets[Context::frameIndex()], &data, sizeof(data));
     }
 
-    const T& operator * (int) const {
+    const T& operator * () const {
       return data;
     }
